requirements.txt
flask==3.0.3
pandas>=2.2.3
scikit-learn>=1.5.2
numpy>=2.1.0
matplotlib>=3.10.0
seaborn==0.13.2
openpyxl==3.1.5
python-dotenv==1.0.1
werkzeug==3.0.4
gunicorn==22.0.0